requirements.txt
curses-menu==0.5.0
PyYAML==5.4.1
